requirements.txt
Flask
Flask-SocketIO
gevent
gunicorn
google-generativeai
python-dotenv